+++ CHANGELOG.md
@@ -19,6 +19,9 @@
 * Updated to Gradle 7.2.
 * Updated to Android Gradle Plugin 7.1.0-beta02.
 * Updated to NDK 23.1.7779620.
+* Updated to Android targetSdk 31.
+* Updated to Android compileSdk 31. 
+* Updated to Android Build Tools 31.0.0.
 
 
 ## 0.7.0-SNAPSHOT (YYYY-MM-DD)